requirements.txt
Django==1.10
psycopg2 >= 2.5.4
djangorestframework==3.5
mysqlclient==1.3.13
#djangorestframework-jwt==1.8.0
MySQL-python==1.2.5